% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents.R
\name{agent_listings}
\alias{agent_listings}
\title{Retrieves the set of listings associated with a specified agent from one of Retsly's MLS datasets.}
\usage{
agent_listings(opts = retsly_query(), dataset = "test", agent_id)
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}

\item{dataset}{Vendor to get the listings from.  See \code{list_datasets} for a list of vendors.}

\item{agent_id}{Agent ID.}
}
\value{
data.frame with listings for each agent.
}
\description{
Retrieves the set of listings associated with a specified agent from one of Retsly's MLS datasets.
}

