% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents.R
\name{agents}
\alias{agents}
\title{Retrieves the set of agents from one of Retsly's MLS datasets.}
\usage{
agents(opts = retsly_query(), dataset = "test", agent_id = NULL)
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}

\item{dataset}{Vendor to get the listings from.  See \code{list_datasets} for a list of vendors.}

\item{agent_id}{Agent ID.}
}
\value{
data.frame of agent information including name, contact info, office ID, and other info.
}
\description{
Retrieves the set of agents from one of Retsly's MLS datasets.
}

