% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{applications}
\alias{applications}
\title{Retrieves the set of applications associated with a user id token.}
\usage{
applications(opts = retsly_query())
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}
}
\value{
list of information including approved vendors (available datasets)
}
\description{
Retrieves the set of applications associated with a user id token.
}

