% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{assessments}
\alias{assessments}
\title{Retrieves the set of public assessments from one of Retsly's datasets.}
\usage{
assessments(opts = retsly_query(), dataset = "pub")
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}

\item{dataset}{Vendor to get the listings from.  See \code{list_datasets} for a list of vendors.}
}
\value{
data.frame with id, fips, location, taxID, assessed value and other information.
}
\description{
Retrieves the set of public assessments from one of Retsly's datasets.
}

