% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{list_datasets}
\alias{list_datasets}
\title{A listing of all the available datasets available on Retsly}
\usage{
list_datasets()
}
\value{
a vector of valid dataset names
}
\description{
A listing of all the available datasets available on Retsly
}

