% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listings.R
\name{listings}
\alias{listings}
\title{Retrieves the set of real estate listings from one of Retsly's MLS datasets.}
\usage{
listings(opts = retsly_query(), dataset = "test", list_id = NULL)
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}

\item{dataset}{Vendor to get the listings from.  See \code{list_datasets} for a list of vendors.}

\item{list_id}{Listing ID for a specific listing.}
}
\value{
data.frame of listing information address, price, agent and other information.
}
\description{
Retrieves the set of real estate listings from one of Retsly's MLS datasets.
}

