% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listings.R
\name{listings_office}
\alias{listings_office}
\title{Retrieves the office associated with a specific real estate listing.}
\usage{
listings_office(opts = retsly_query(), dataset = "test", list_id)
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}

\item{dataset}{Vendor to get the listings from.  See \code{list_datasets} for a list of vendors.}

\item{list_id}{Listing ID for a specific listing.}
}
\value{
List of information including officeID, name, address and other information.
}
\description{
Retrieves the office associated with a specific real estate listing.
}

