% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{media}
\alias{media}
\title{Retrieves the collection of media for a set of listings.}
\usage{
media(opts = retsly_query(), dataset = "test", media_id = NULL)
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}

\item{dataset}{Vendor to get the listings from.  See \code{list_datasets} for a list of vendors.}

\item{media_id}{media ID.}
}
\value{
data.frame of id, url and mimeType.
}
\description{
Retrieves the collection of media for a set of listings.
}

