% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offices.R
\name{offices}
\alias{offices}
\title{Retrieves the set of real estate offices from one of Retsly's MLS datasets.}
\usage{
offices(opts = retsly_query(), dataset = "test", office_id = NULL)
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}

\item{dataset}{Vendor to get the listings from.  See \code{list_datasets} for a list of vendors.}

\item{agent_id}{Office ID.}
}
\value{
data.frame of office information including name, address, officeID and other information.
}
\description{
Retrieves the set of real estate offices from one of Retsly's MLS datasets.
}

