% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offices.R
\name{offices_listings}
\alias{offices_listings}
\title{Retrieves the set of listings from a specific office for one of Retsly's MLS datasets.}
\usage{
offices_listings(opts = retsly_query(), dataset = "test", office_id)
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}

\item{dataset}{Vendor to get the listings from.  See \code{list_datasets} for a list of vendors.}

\item{office_id}{Office ID.}
}
\value{
data.frame of listing information including address, listingID and other information.
}
\description{
Retrieves the set of listings from a specific office for one of Retsly's MLS datasets.
}

