% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcels.R
\name{parcels}
\alias{parcels}
\title{Retrieves the set of parcels from Retsly's datasets.}
\usage{
parcels(opts = retsly_query(), dataset = "pub")
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}

\item{dataset}{Vendor to get the listings from.  See \code{list_datasets} for a list of vendors.}
}
\value{
data.frame of parcel information including id, land use description, acres and other information.
}
\description{
Retrieves the set of parcels from Retsly's datasets.
}

