% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{retsly_query}
\alias{retsly_query}
\title{Retsly API query parameters}
\usage{
retsly_query(access_token = "790d218426a115416a7f69371094d28e",
  vendor = NULL, offset = NULL, limit = 100, sortBy = NULL,
  order = NULL, near = NULL, radius = NULL, total = NULL, ...)
}
\arguments{
\item{access_token}{Retsly API access token.}

\item{vendor}{Vendor to get the listings from.  See \code{list_datasets} for a list of vendors.}

\item{offset}{Skips the number of results.}

\item{limit}{Limits the size of the result set.}

\item{sortBy}{Response field to sort by.}

\item{order}{Order by "asc" or "desc".}

\item{near}{Coord or location either "lat,long" or "San Diego".}

\item{radius}{Search radius.  For example "10mi", "10km" or degrees (if no units).}

\item{...}{Additional query parameters, see \code{valid_query}}
}
\value{
Named list of API query parameters.
}
\description{
Retsly API query parameters
}

