% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transactions.R
\name{transactions}
\alias{transactions}
\title{Retrieves the set of transactions from one of Retsly's datasets.}
\usage{
transactions(opts = retsly_query(), dataset = "pub")
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}

\item{dataset}{Vendor to get the listings from.  See \code{list_datasets} for a list of vendors.}
}
\value{
data.frame of transactions including state, documentType, lenderName, parcels and other information.
}
\description{
Retrieves the set of transactions from one of Retsly's datasets.
}

