% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users}
\alias{users}
\title{Retrieve a user associated with a user token.}
\usage{
users(opts = retsly_query())
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}
}
\value{
data.frame
}
\description{
Retrieve a user associated with a user token.
}

