% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{valid_query}
\alias{valid_query}
\title{Is the parameter a valid query parameter for the Retsly API?}
\usage{
valid_query(query)
}
\arguments{
\item{Retsly}{query}
}
\value{
TRUE if valid operator otherwise FALSE
}
\description{
Is the parameter a valid query parameter for the Retsly API?
}

