% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendors.R
\name{vendors}
\alias{vendors}
\title{Retrieves the set of all vendors available.}
\usage{
vendors(opts = retsly_query(), vendor_id = NULL)
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}

\item{vendor_id}{Specific vendor ID.}
}
\value{
data.frame of all vendors including name, id, number of listings and other information.
}
\description{
Retrieves the set of all vendors available.
}

