% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendors.R
\name{vendors_approved}
\alias{vendors_approved}
\title{Retrieves the set of vendors approved for a user.}
\usage{
vendors_approved(opts = retsly_query())
}
\arguments{
\item{opts}{Request query parameters, see \code{retsly_query}.}
}
\value{
data.frame of approved vendors including name, id, number of listings and other information.
}
\description{
Retrieves the set of vendors approved for a user.
}

